+++ tests/srcs/strlcat_test.c
@@ -18,20 +18,20 @@ void strlcat_test(const char *s1, const char *s2, size_t dstsize, size_t catsize
 	printf("src: %s\ndest: %s\nn: %ld\n", src, dst, catsize);
 	printf("##MINE##\nRV: %ld\nword: %s\n", ft_strlcat(dst, src, catsize), dst);
 	printf("String: ");
-	for (int i = 0; i < ft_strlen(dst) + 1; i++)
+	for (size_t i = 0; i < ft_strlen(dst) + 1; i++)
 		printf("%c ", dst[i]);
 	putstr("");
 	printf("Int: ");
-	for (int i = 0; i < ft_strlen(dst) + 1; i++)
+	for (size_t i = 0; i < ft_strlen(dst) + 1; i++)
 		printf("%d ", dst[i]);
 	putstr("\n");
 	printf("##LIB##\nRV: %ld\nword: %s\n", strlcat(libdst, libsrc, catsize), libdst);
 	printf("String: ");
-	for (int i = 0; i < ft_strlen(libdst) + 1; i++)
+	for (size_t i = 0; i < ft_strlen(libdst) + 1; i++)
 		printf("%c ", libdst[i]);
 	putstr("");
 	printf("Int: ");
-	for (int i = 0; i < ft_strlen(libdst) + 1; i++)
+	for (size_t i = 0; i < ft_strlen(libdst) + 1; i++)
 		printf("%d ", libdst[i]);
 	putstr("\n\n");
 	free(dst);
